requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
defusedxml==0.5.0         # via jira
django-govuk-forms==0.6
django-govuk-template==0.7
django==2.0.7
idna==2.7                 # via requests
jira==1.0.15
libsass==0.14.5
oauthlib==2.1.0           # via requests-oauthlib
pbr==4.1.0                # via jira
pytz==2018.5              # via django
raven==6.9.0
requests-oauthlib==1.0.0  # via jira
requests-toolbelt==0.8.0  # via jira
requests==2.19.1          # via jira, requests-oauthlib, requests-toolbelt
six==1.11.0               # via jira, libsass
urllib3==1.23             # via requests
